+++ test/smokes/cpplint/default_rules/src/foo/utils.h
@@ -1,0 +1,6 @@
+/* Copyright [2020] John Doe */
+#ifndef UTILS_H_
+#define UTILS_H_
+#include <vector>
+#include <string>
+#endif  // UTILS_H_