+++ test/smokes/cpplint/default_rules_when_cfg_exists/src/bar/test1.cpp
@@ -1,0 +1,7 @@
+#include <cstdio>
+
+int main()
+{
+  char* str;
+  printf(str);
+}