+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - **ESLint** Add `eslint-plugin-vue` to default package list [#1521](https://github.com/sider/runners/pull/1521)
 - **RuboCop** Update optional gem list [#1506](https://github.com/sider/runners/pull/1506)
 - Delete environment variables after used [#1520](https://github.com/sider/runners/pull/1520)
+- **cpplint** Provide Sider's recommended ruleset [#1505](https://github.com/sider/runners/pull/1505)
 
 ## 0.35.0